Python Code for Paper
This code can be accessed via JupyterLab and shows the analysis conduced to to analyse the feasibility of implementing an web application to assist Microfinance groups' organisation and management. It involves data preprocessing and cleaning, some data exploration, resampling of the training data. The prediction is assessed based on two models, logistic regression with Hyperparameter tuning and logistic regression based on P-values. Confusion matrices are created to assess which model performs best.